nr pdcp: add --pdcp-drop option to drop packets in PDCP instead of RLC
This only works in monolithic. F1 split to be done later, it's a bit more
complicated.

And actually, the drop in RLC is still there, but it should not happen.
Or much less.

The logic around rlc_data_queue_occupancy() is not satisfying. It is
done globally for all the UEs so we may underestimate the real free
space in RLC for a given UE, and thus reject more packets than
necessary.